Fixed library crash
A variable in BlockLayered previously overshadowed a variable in Block in deobfuscated environements, causing uncaught runtime errors under certain conditions. This only presented an issue when attempting to use Minestuck as a library in a dev environement, rather than using it as a mod or as a project. The name of the variable has been changed, and the overshadowing no longer occurs, fixing the issue.