Ender Fluid
Ender, the liquid that defines the oceans of up-and-coming End Lands, is unique in several ways. Firstly, it is completely opaque to the player, being impossible to look through, but it does not block light levels passing through it. Secondly, It only flows half as far east/west as it does north/south. Thirdly, it self-replicates like water under the new moon. Fourthly, it does not flow during the full moon.