Alchemiter splash particles and sound changes (#472)
* added redsplat, redpop, and plasma particle

includes item classes, and adds functionality to alchemiter

* extracted one line code

* fixed error

* size change, and height reorientation

increased the z height for the particle spawn so that it looks a bit better and shrunk the particles to fit the alchemiter a bit

* changed the particle classes

modified sizes and removed some of the less needed comments. Also increased the z value by 0.2

* added subtitles, and a sound to the action

* added sounds to transportalizer, and totem lathe

* PR review update, alpha particles

updated following the first round of PR reviews. riotmode added alpha to the particles, it's not enough so they'll be changed in a future commit

* Changelog change

* changed changelog

* git review update

* Update src/main/java/com/mraof/minestuck/blockentity/machine/AlchemiterBlockEntity.java

Co-authored-by: Dweblenod <63124354+Dweblenod@users.noreply.github.com>

* Update src/main/java/com/mraof/minestuck/client/particles/PlasmaParticle.java

Co-authored-by: Dweblenod <63124354+Dweblenod@users.noreply.github.com>

* Update src/main/resources/assets/minestuck/particles/plasma.json

Co-authored-by: Dweblenod <63124354+Dweblenod@users.noreply.github.com>

* removed plasma particle 6

---------

Co-authored-by: Dweblenod <63124354+Dweblenod@users.noreply.github.com>