Try using a spring-like approach to enforcing editmode border
The previous approach of removing outward velocity wasn't working because it wasn't happening between acceleration and movement.
An alternative I didn't try was adding a collision box akin to the vanilla world border. Problem is that there was no hook by which to do so, and I was not prepared to create my own for this.