Carved Bush (#565)
Have a Carved Bush. As Aki is working on the overall Wood terrain
glow-up, I didn't put the bush in the changelog. Presumably, the
"WOOD_TERRAIN_BLOCKS" tag will also be updated to reflect whatever the
terrain blocks are. I've already provided the updates for the tag, to
accommodate fixes Aki is working on for the Wooden Cactus. (Real PR)
(NOT CLICKBAIT)

Texture by Glubtier.

---------

Co-authored-by: Dweblenod <63124354+Dweblenod@users.noreply.github.com>