Skaia blocks class and item-block registry helper (#581)
Moves various blocks and block items to a new class `SkaiaBlocks`. Also
includes a new class `ItemBlockPair` for registering and storing a block
and its block item together.

Moves datagen for skaia blocks to new class `SkaiaBlocksData`. I decided
to not include skaia worldgen datagen in this because of how little
there is.