Skaianet and Sburb connection cleanup (#545)
Separates `SburbConnection` into two groups of data:
- `ActiveConnection`, which has references to the computers used in the
connection. Relevant for entering editmode.
- `SburbPlayerData` for any data that is more closely tied to the
player/that should not disappear when the computer connection is closed.
Includes entry status and primary player connections.

Removes some unused, poorly used or low-value features to simplify
related code:
- Locking a sburb connection or predefined data to a specific session.
Removed as it wasn't really used at this point.
- Locking a session so that no additional players could join a session.
Was meant to be part of the sburb session command (the predecessor to
sburb predefine which had a focus on predefining which players belonged
to the same session, unlike the simplified predefine that exists today).
- Session name, which was used to identify sessions with commands. Was
part of the sburb session command, but ceased being used with the
simplified sburb predefine.
- The hard requirement that two identical titles shouldn't exist in the
same session. While it does still make sense to avoid duplicate titles,
I don't think it is worth the code to make it a hard requirement.
- Session size limit of 144 players. It was added to support the hard
requirement of no duplicate titles, so it doesn't make sense to keep
when removing said hard requirement.

The code surrounding connections is also becoming less object-oriented
and more data-oriented.
Spaghetti code is being untangled, redirects are shortened.
`SkaianetHandler` got split in parts, predefine data was moved out of
sessions.
The internal api for connections was updated to be more reliable to be
used by code outside the `skaianet` package. Some code that didn't quite
belong in the package gets moved closer to their caller.