Holopad-focused code cleanup (#623)
Removes the unused hologram entity (the current version of the holopad
functions through its block entity renderer) and some code associated
with it, and performs some general code cleanup on both the block and
block entity.

Changed the holopad to display a generic object instead of the
contained item when the card isn't punched, instead of just when the
card is empty. This is to make the holopad behave as if it was
specifically reading the item off of the punched holes of the card.

Fixed collision box.