Fix and remove smaller todo comments (#649)
Goes through current todo and fixme comments, fixing or implementing a
selection of them based on how small and straightforward I considered
them to be. Also removed some comments that either did not need an
action, was already solved, or were under-specified feature suggestions.
I've also done some code cleanup related to code with a todo, even when
the cleanup is not related to the todo itself.