Draw skybox veil meteors in one rendering batch for better performance (#683)
Using drawSprite() will draw it one sprite at a time, which means one
rendering call for each of the 1500 meteors. The new solution (similar
to what is used in other places of skybox rendering) is to add all
meteors to the same buffer builder and then render all of them in one
call.