Set teleport pos in the compound tag before loading new entity (#684)
This fixes a compatibility issue with Create where certain contraptions
(such as windmills) would consistently break during entry.

The fix is to ensure that the new position is stored in the nbt data for
the new entity before invoking `entity.load()`. By doing this, the
entity is able to load the correct position for the corresponding
contraption block.

Note that this just fixes the most obvious issue. Machines can still
break in particular ways if for example only the block or only the
entity is transferred during entry. But these type of edge cases are not
the focus of this PR.